# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Work In Progress",
        "description": "Dokumentasi API untuk proses inspeksi, SWA, dan validasi",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/getuserbygroup/{groupUser}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get users by group",
                "operationId": "getUserByGroup",
                "parameters": [
                    {
                        "name": "groupUser",
                        "in": "path",
                        "description": "The name of the user group",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Safety Advisor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User get data successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Users not found or already deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User get data not found or already deleted"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get all users",
                "operationId": "getAllUsers",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of templates per page",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User get data successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "UserId": {
                                                        "type": "string",
                                                        "format": "uuid"
                                                    },
                                                    "Username": {
                                                        "type": "string",
                                                        "example": "ekoagung"
                                                    },
                                                    "UserGroup": {
                                                        "type": "string",
                                                        "example": "Safety Advisor"
                                                    },
                                                    "user_detail": {
                                                        "properties": {
                                                            "Name": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/{userId}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user detail by ID",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "d547dfc0-5177-4b64-a491-3e41f4c3a312"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User detail retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User get data successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "UserId": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "Name": {
                                                    "type": "string"
                                                },
                                                "Email": {
                                                    "type": "string"
                                                },
                                                "Address": {
                                                    "type": "string"
                                                },
                                                "City": {
                                                    "type": "string"
                                                },
                                                "Phone": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user details",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "Email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "Address": {
                                        "type": "string",
                                        "example": "Jl. Merdeka 123"
                                    },
                                    "City": {
                                        "type": "string",
                                        "example": "Jakarta"
                                    },
                                    "Phone": {
                                        "type": "string",
                                        "example": "081234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user by ID (Admin only)",
                "operationId": "fb8b15b1ba4e99d763455d403adc96ec",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "UserId": {
                                                    "type": "string"
                                                },
                                                "Username": {
                                                    "type": "string"
                                                },
                                                "UserGroup": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Only Administrator can delete users"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/{userId}/change-password": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Change user password",
                "operationId": "changePassword",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "OldPassword",
                                    "Password"
                                ],
                                "properties": {
                                    "OldPassword": {
                                        "type": "string",
                                        "example": "oldpassword123"
                                    },
                                    "Password": {
                                        "type": "string",
                                        "example": "newpassword456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password changed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Password changed successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/user/groups": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get all user groups (except Administrator)",
                "operationId": "0e058aed23e060be4e3a711c571c264e",
                "responses": {
                    "200": {
                        "description": "List of user groups",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "UserGroup": {
                                                        "type": "string",
                                                        "example": "Supervisor"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a bearer token to access this API",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}